Cleanup based on review comments
