+++ Modix.Services/Utilities/FormatUtilities.cs
@@ -109,5 +109,19 @@ namespace Modix.Services.Utilities
 
             return $"This user has {formatted}";
         }
+
+        public static string Sanitize(string text)
+        {
+            var everyoneSanitized = _everyoneMentionRegex.Replace(text, _sanitizedEveryoneMention);
+            var hereSanitized = _hereMentionRegex.Replace(everyoneSanitized, _sanitizedHereMention);
+
+            return hereSanitized;
+        }
+
+        private static readonly Regex _everyoneMentionRegex = new Regex("@everyone", RegexOptions.Compiled);
+        private const string _sanitizedEveryoneMention = "@\x200beveryone";
+
+        private static readonly Regex _hereMentionRegex = new Regex("@here", RegexOptions.Compiled);
+        private const string _sanitizedHereMention = "@\x200bhere";
     }
 }
